+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [Unreleased]
+
+### Added
+
+- gremlin: support for `halt` action.
+
+
+
 ## [1.5.0] - 2019-03-29
 
 ### Added